Better file separation | Using ft_message_error
