+++ Makefile
@@ -14,6 +14,7 @@ SRCS = 	main.c \
 		$(SRC_DIR)/ft_execution/ft_render/ft_floor_ceiling_render.c \
 		$(SRC_DIR)/ft_execution/ft_render/ft_raycasting.c \
 		$(SRC_DIR)/ft_execution/ft_movement/ft_movement.c \
+		$(SRC_DIR)/ft_execution/ft_init.c \
 		$(SRC_DIR)/ft_execution/ft_window.c \
 		$(SRC_DIR)/ft_execution/outils/outils.c \
 		$(SRC_DIR)/ft_parsing/ft_parsing.c \
