+++ srcs/ft_execution/ft_render/ft_load_textures.c
@@ -6,7 +6,7 @@
 /*   By: ahors <ahors@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/30 14:10:03 by ahors             #+#    #+#             */
-/*   Updated: 2024/10/15 11:04:49 by ahors            ###   ########.fr       */
+/*   Updated: 2024/10/15 12:41:09 by ahors            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
